Add support for including git commit details in all messages.
Added new `git_info` segment to the opinionated messages, which states:

```
Changes by ${GIT_COMMITTER}.
- [${GIT_SHORT_REF}] ${GIT_COMMIT_MESSAGE}
```

This is appended right after the "normal" message (on the first line).
The commit short sha and message are added on the next line.

See the updated `README.md` for full details and configuration options.